+++ mantle/Cargo.toml
@@ -10,9 +10,15 @@ readme = "README.md"
 keywords = ["mantle"]
 
 [dependencies]
+blockchain-traits = { version = "0.1", path = "../blockchain-traits" }
+cfg-if = "0.1"
 failure = "0.1"
 mantle-macros = { version = "0.2", path = "../mantle-macros" }
 mantle-types = { version = "0.2", path = "../mantle-types" }
 serde = { version = "1.0", features = ["derive"] }
 serde_cbor = "0.9"
 tiny-keccak = "1.4"
+
+[target.'cfg(target_os = "wasi")'.dependencies]
+hex = "0.3"
+libc = "0.2"
